reduce projs
When encountering a projection, we weak-head reduce its subject. If we obtain a
constructor, we reduce the projection immediately (without backtracking).
Otherwise and if the reduced subject is not ground, we try solving the
canonical structure problem after refolding the other term to its state
at the start of the unification.